SDK-2534: Move Identity Profile responses and associated classes to their own package
