+++ yoti-sdk-api/src/main/java/com/yoti/api/client/docs/session/retrieve/identityprofile/IdentityProfilePreviewResponse.java
@@ -1,4 +1,6 @@
-package com.yoti.api.client.docs.session.retrieve;
+package com.yoti.api.client.docs.session.retrieve.identityprofile;
+
+import com.yoti.api.client.docs.session.retrieve.MediaResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
