+++ yoti-sdk-api/src/main/java/com/yoti/api/client/docs/session/retrieve/identityprofile/IdentityProfileReportResponse.java
@@ -1,6 +1,8 @@
-package com.yoti.api.client.docs.session.retrieve;
+package com.yoti.api.client.docs.session.retrieve.identityprofile;
 
 import java.util.List;
+
+import com.yoti.api.client.docs.session.retrieve.MediaResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
