+++ yoti-sdk-api/src/main/java/com/yoti/api/client/docs/session/retrieve/identityprofile/IdentityProfileResponse.java
@@ -1,4 +1,4 @@
-package com.yoti.api.client.docs.session.retrieve;
+package com.yoti.api.client.docs.session.retrieve.identityprofile;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
